refactor(qwen35): dedupe the CUDA thread binding into cublas_thread.rs
The scheduler and the TP worker carried byte-identical CUDA/cuBLAS thread
bindings (guard + set_device + bind_to_thread + cublas_init); the worker
split moved one of the two copies, making the duplication visible. Both
now call crate::cublas_thread::bind_model_thread with a role tag (the
failure messages are unchanged); the scheduler keeps its
tune_decode_gemm_algos step at its own call site. Behavior unchanged.
