+++ pegainfer-qwen35/src/tp_executor/worker.rs
@@ -2,6 +2,7 @@
 //! gating, the NCCL startup watchdog, and the per-rank command loop.
 
 use super::*;
+use crate::cublas_thread::CublasThreadGuard;
 
 const TP_NCCL_STARTUP_TIMEOUT: std::time::Duration = std::time::Duration::from_secs(60);
 const TP_WORKER_SHUTDOWN_TIMEOUT: std::time::Duration = std::time::Duration::from_secs(30);
@@ -219,7 +220,7 @@ impl TpWorkerPrepared {
         max_prefill_tokens: usize,
         graph_enabled: bool,
     ) -> Result<(Self, usize)> {
-        let cublas_guard = bind_worker_thread(&model)?;
+        let cublas_guard = crate::cublas_thread::bind_model_thread(&model, "TP worker")?;
         let (free_bytes, total_bytes) = model
             .device_ctx()
             .ctx
@@ -1109,37 +1110,6 @@ fn sample_decode_rows(
             }
         })
         .collect())
-}
-
-struct CublasThreadGuard;
-
-impl Drop for CublasThreadGuard {
-    fn drop(&mut self) {
-        unsafe {
-            crate::ffi::cublas_destroy();
-        }
-    }
-}
-
-fn bind_worker_thread(model: &Qwen35Model) -> Result<CublasThreadGuard> {
-    let ctx = model.device_ctx();
-    unsafe {
-        let err = crate::ffi::cuda_set_device(ctx.device_ordinal as i32);
-        if err != 0 {
-            return Err(anyhow::anyhow!(
-                "Failed to set CUDA device {} on Qwen3.5 TP worker thread: cudaError={}",
-                ctx.device_ordinal,
-                err
-            ));
-        }
-    }
-    ctx.ctx.bind_to_thread().map_err(|e| {
-        anyhow::anyhow!("Failed to bind CUDA context to Qwen3.5 TP worker thread: {e}")
-    })?;
-    unsafe {
-        crate::ffi::cublas_init();
-    }
-    Ok(CublasThreadGuard)
 }
 
 #[cfg(test)]
